ekfac_tests: Use float32 in ground truth computation
Match the float32 precision used in the main EKFAC implementation:
- GroundTruthCovarianceCollector: Convert activations/gradients to
  float32 before computing covariances
- GroundTruthAmortizedLambdaCollector: Convert to float32 for einsum
- GroundTruthNonAmortizedLambdaCollector: Convert to float32 for einsum
- compute_eigenvectors_step: Store eigenvectors in float32

Add jaxtyping annotations (Float, Float32) for precise documentation
of tensor shapes and dtypes throughout the ground truth collectors.

+++ tests/ekfac_tests/compute_ekfac_ground_truth.py
@@ -586,8 +586,9 @@ def compute_eigenvectors_step(
             f"{name}: eigenvectors_a.sum()={eigenvectors_a.sum()}, "
             f"eigenvectors_g.sum()={eigenvectors_g.sum()}"
         )
-        eigenvectors_activations[name] = eigenvectors_a.to(dtype=dtype).contiguous()
-        eigenvectors_gradients[name] = eigenvectors_g.to(dtype=dtype).contiguous()
+        # Store in float32 to match EKFAC implementation
+        eigenvectors_activations[name] = eigenvectors_a.to(torch.float32).contiguous()
+        eigenvectors_gradients[name] = eigenvectors_g.to(torch.float32).contiguous()
 
     save_file(
         eigenvectors_activations,
